Split context.py in context.py and context_stack.py, added code from feedburner into calendar, Constants now in constants.py, Calendar in calendar.py
