+++ tests/TestAlternativeAbbreviations.py
@@ -4,7 +4,7 @@ import sys
 import time
 import datetime
 import parsedatetime as pdt
-from parsedatetime.pdt_locales import get_icu
+from parsedatetime.locales import get_icu
 from . import utils
 
 if sys.version_info < (2, 7):
