+++ tests/TestContext.py
@@ -1,12 +1,12 @@
 # -*- coding: utf-8 -*-
 """
-Test pdtContext
+Test Context
 """
 
 import sys
 import time
 import parsedatetime as pdt
-from parsedatetime.context import pdtContext
+from parsedatetime.context import Context
 
 if sys.version_info < (2, 7):
     import unittest2 as unittest
@@ -23,26 +23,26 @@ class test(unittest.TestCase):
 
     def testContext(self):
         self.assertEqual(self.cal.parse('5 min from now')[1],
-                         pdtContext(pdtContext.ACU_MIN | pdtContext.ACU_NOW))
+                         Context(Context.ACU_MIN | Context.ACU_NOW))
         self.assertEqual(self.cal.parse('5 min from now',
                                         version=pdt.VERSION_FLAG_STYLE)[1], 2)
         self.assertEqual(self.cal.parse('7/11/2015')[1],
-                         pdtContext(pdtContext.ACU_YEAR |
-                                    pdtContext.ACU_MONTH | pdtContext.ACU_DAY))
+                         Context(Context.ACU_YEAR |
+                                    Context.ACU_MONTH | Context.ACU_DAY))
         self.assertEqual(self.cal.parse('7/11/2015',
                                         version=pdt.VERSION_FLAG_STYLE)[1], 1)
         self.assertEqual(self.cal.parse('14/32/2015')[1],
-                         pdtContext(0))
+                         Context(0))
         self.assertEqual(self.cal.parse('25:23')[1],
-                         pdtContext())
+                         Context())
 
     def testSources(self):
         self.assertEqual(self.cal.parse('afternoon 5pm')[1],
-                         pdtContext(pdtContext.ACU_HALFDAY |
-                                    pdtContext.ACU_HOUR))
+                         Context(Context.ACU_HALFDAY |
+                                    Context.ACU_HOUR))
 
         self.assertEqual(self.cal.parse('morning')[1],
-                         pdtContext(pdtContext.ACU_HALFDAY))
+                         Context(Context.ACU_HALFDAY))
 
         self.assertEqual(self.cal.parse('night', version=1)[1], 2)
